libleptonrenderer: Fix several warnings about comparison between integers.
This fixes the following warnings when compiling with g++:
  "warning: comparison between signed and unsigned integer
  expressions [-Wsign-compare]"

Where possible, the type of iterator was changed, otherwise
explicit casts were used.

+++ libleptonrenderer/edacairo.c
@@ -61,7 +61,7 @@ eda_cairo_set_source_color (cairo_t *cr, int color, GArray *map)
 
   g_return_if_fail (color >= 0);
   g_return_if_fail (map != NULL);
-  g_return_if_fail ((color >= 0) && (map->len > color));
+  g_return_if_fail ((color >= 0) && ((int) map->len > color));
 
   c = g_array_index (map, GedaColor, color);
